requirements.txt
gTTS==2.4.0
num2words==0.5.12
Pillow==10.0.1
playsound==1.2.2
pygame==2.5.2